// web-ui/client/gen/sales/opportunity.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */
import type { StageEnum } from './stageEnum';

export interface Opportunity {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  account: number;
  readonly account_name: string;
  /** @nullable */
  contact?: number | null;
  readonly contact_name: string;
  /** @pattern ^-?\d{0,10}(?:\.\d{0,2})?$ */
  amount: string;
  stage?: StageEnum;
  /** @nullable */
  close_date?: string | null;
  /** @nullable */
  owner?: number | null;
  readonly owner_name: string;
  readonly created_at: string;
}